+++ gui/Base/MainView.py
@@ -31,9 +31,9 @@ This module defines the following classes:
 - NewCaseDialogView
 - MainView
 
-    @copyright: 1998-2017 EDF S.A., France
+    @copyright: 1998-2018 EDF S.A., France
     @author: U{EDF<mailto:saturne-support@edf.fr>}
-    @license: GNU GPL v2, see COPYING for details.
+    @license: GNU GPL v2 or later, see COPYING for details.
 """
 
 #-------------------------------------------------------------------------------